+++ HotelManagerSystem.Models/Entities/Hotel.cs
@@ -10,7 +10,7 @@ namespace HotelManagerSystem.Models.Entities
         public User User { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public int ReviewStars { get; set; }
+        public decimal ReviewStars { get; set; }
         public bool IsOne { get; set; }
         public string CheckingAccount { get; set; }
         public int FilialCount { get; set; }
